Footnote tool - initial upload
This is the best I could come up with to distinguish footnote references from other numbers in the text; which is a very difficult thing to program. Instruction: You select the first footnote reference and the algorithm will count up from it, where the line gets interrupted you select the next footnote reference again.